+++ packages/react-nft-checkout/src/components/PaymentTokensList/PaymentTokensList.tsx
@@ -12,6 +12,10 @@ import { useEffect, useState } from 'react'
 import { ErrorText, LoadingIndicator } from '@/components'
 import { useDappContext } from '@/hooks'
 
+/**
+ * @description A list of the tokens in the user's selected wallet
+ * @group Components
+ */
 const PaymentTokensList = () => {
   const {
     selectedChain,