+++ packages/react-nft-checkout/src/components/PaymentWallets/PaymentWallets.tsx
@@ -31,6 +31,10 @@ const PROVIDERS_DATA = [
   },
 ]
 
+/**
+ * @description A list of the wallets detected in the user's browser
+ * @group Components
+ */
 const PaymentWallets = () => {
   const { provider, setSelectedProviderProxy, createProviderError } =
     useDappContext()