+++ CHANGELOG.md
@@ -6,10 +6,8 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
 ## [Unreleased]
+## [1.5.0-rc.5] - 2023-03-30
 ### Added
-- `@rarimo/react-nft-checkout`: React implementation of `@rarimo/nft-checkout` package
-- `@rarimo/react-nft-checkout`: Exported UI component `RarimoPayButon`
-- `@rarimo/react-nft-checkout`: Exported context provider `DappContextProvider` with core methods for NFT chekout
 - `@rarimo/nft-checkout`: Selectable token to swap
 - `@rarimo/nft-checkout`: Disabled native token and wrrapped native tokens
 
@@ -18,6 +16,37 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - `@rarimo/provider`: Automatic change chain
 - `@rarimo/provider`: Load supported tokens
 
+## [1.5.0-rc.4] - 2023-03-24
+### Fixed
+- `@rarimo/react-nft-checkout`: yarn.lock `jsbi` version
+
+## [1.5.0-rc.3] - 2023-03-24
+### Fixed
+- `@rarimo/react-nft-checkout`: Trying to fix `jsbi` dependency issue, added `@uniswap/router-sdk` alias to the vite config to resolve esm module
+
+## [1.5.0-rc.2] - 2023-03-24
+### Added
+- `@rarimo/react-nft-checkout`: `jsbi` dependency to cover `@uniswap/smart-order-router` peer-dependency
+
+## [1.5.0-rc.1] - 2023-03-24
+### Added
+- `@rarimo/react-nft-checkout`: `NODE_ENV=production` to the build script to force Vite to resolve production modules
+- `@rarimo/react-provider`: `NODE_ENV=production` to the build script force Vite to resolve production modules
+- `root`: `apply-version` Script to easier bump version
+
+### Fixed
+- `@rarimo/react-nft-checkout`: Types path
+- `@rarimo/react-nft-checkout`: npmignore
+- `@rarimo/react-provider`: npmignore
+
+## [1.5.0-rc.0] - 2023-03-23
+### Added
+- `@rarimo/react-nft-checkout`: React implementation of `@rarimo/nft-checkout` package
+- `@rarimo/react-nft-checkout`: Exported UI component `RarimoPayButon`
+- `@rarimo/react-nft-checkout`: Exported context provider `DappContextProvider` with core methods for NFT checkout
+
+### Fixed
+- `@rarimo/react-provider`: [near-api-js](https://github.com/vitejs/vite/issues/9703) vite build issue
 
 ## [1.4.0] - 2023-03-13
 ### Added
@@ -104,7 +133,13 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ### Under the hood changes
 - Initiated repo with `@rarimo/provider` and `@rarimo/nft-checkout` packages
 
-[Unreleased]: https://github.com/rarimo/js-sdk/compare/1.4.0...HEAD
+[Unreleased]: https://github.com/rarimo/js-sdk/compare/1.5.0-rc.5...HEAD
+[1.5.0-rc.5]: https://github.com/rarimo/js-sdk/compare/1.5.0-rc.4...1.5.0-rc.5
+[1.5.0-rc.4]: https://github.com/rarimo/js-sdk/compare/1.5.0-rc.3...1.5.0-rc.4
+[1.5.0-rc.3]: https://github.com/rarimo/js-sdk/compare/1.5.0-rc.2...1.5.0-rc.3
+[1.5.0-rc.2]: https://github.com/rarimo/js-sdk/compare/1.5.0-rc.1...1.5.0-rc.2
+[1.5.0-rc.1]: https://github.com/rarimo/js-sdk/compare/1.5.0-rc.0...1.5.0-rc.1
+[1.5.0-rc.0]: https://github.com/rarimo/js-sdk/compare/1.4.0...1.5.0-rc.0
 [1.4.0]: https://github.com/rarimo/js-sdk/compare/1.3.0...1.4.0
 [1.3.0]: https://github.com/rarimo/js-sdk/compare/1.2.0...1.3.0
 [1.2.0]: https://github.com/rarimo/js-sdk/compare/1.1.0...1.2.0